feat: replace reducer with slices and create user slice
