+++ src/types/interactions.ts
@@ -1,5 +1,4 @@
 export type Interactions = {
   city: string;
-  sortType: string;
 };
