+++ src/types/user.ts
@@ -1,0 +1,5 @@
+import { AuthorizationStatusType } from './consts';
+
+export type User = {
+  authorizationStatus: AuthorizationStatusType;
+}
